+++ releases.md
@@ -1,3 +1,4 @@
 # pi-clock
 1.1.0
 * option to tick during float animation
+* restore scroll position of settings screen